requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
httpx==0.26.0
